+++ src/structs/mod.rs
@@ -13,6 +13,7 @@
 pub mod dirent;
 pub mod extent;
 pub mod group_desc;
+pub mod htree;
 pub mod inode;
 pub mod superblock;
 pub mod xattr;